requirements.txt
fastapi==0.115.6
feedparser==6.0.11
openai==1.58.1
pydantic==2.10.4
PyMySQL==1.1.1
PyYAML==6.0.2
uvicorn==0.34.0
